api-integration-test/requirements.txt
# API 集成测试依赖
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
hubspot-api-client>=7.0.0
oss2>=2.18.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
